tasklist kadai1

+++ app/Http/Controllers/TasksController.php
@@ -16,13 +16,21 @@ class TasksController extends Controller
         // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::all();
+        if (\Auth::check()) { // 認証済みの場合
+            // MEMO: 認証済みユーザーのオブジェクトを$userに格納している
+            $user = \Auth::user();
+            // タスク一覧を取得
+            // MEMO: $tasksに格納はされているのは、認証済みユーザーのリレーションを元に取得したTaskのレコード
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->get();
 
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            // タスク一覧ビューでそれを表示
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        } else {
+            // Welcomeビューでそれらを表示
+            return view('welcome');
+        }
     }
 
     /**
@@ -32,12 +40,17 @@ class TasksController extends Controller
      */
     public function create()
     {
-        $task = new Task;
+        if (\Auth::check()) { // 認証済みの場合
+            $task = new Task;
 
-        // タスク作成ビューを表示
-        return view('tasks.create', [
-            'task' => $task,
-        ]);
+            // タスク作成ビューを表示
+            return view('tasks.create', [
+                'task' => $task,
+            ]);
+        } else {
+            // トップページへリダイレクトさせる
+            return back();
+        }
     }
 
     /**
@@ -54,11 +67,20 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
     
-        // タスクを作成
-        $task = new task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $task = $user->tasks()->orderBy('created_at', 'desc');
+
+            // タスクを作成
+            $task = new task;
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->user_id = \Auth::id();
+            $task->save();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -72,13 +94,18 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        // idの値でタスクを検索して取得
-        $task = task::findOrFail($id);
+        $task = Task::findOrFail($id);
+        
+        if (\Auth::id() === $task->user_id) { // 認証済みの場合
 
-        // タスク詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+            // タスク詳細ビューでそれを表示
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        } else {
+            // トップページへリダイレクトさせる
+            return redirect('/');
+        }
     }
 
     /**
@@ -89,13 +116,18 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
+        if (\Auth::id() === $task->user_id) { // 認証済みの場合
 
-        // タスク編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+            // タスク編集ビューでそれを表示
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        } else {
+            // トップページへリダイレクトさせる
+            return back();
+        }
     }
 
     /**
@@ -107,21 +139,21 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        
         // バリデーション
         $request->validate([
             'status' => 'required|max:10',
             'content' => 'required|max:255',
         ]);
         
-        // idの値でタスクを検索して取得
-        $task = task::findOrFail($id);
-        // タスクを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
-        // トップページへリダイレクトさせる
+        $task = Task::findOrFail($id);
+        
+        if (\Auth::id() === $task->user_id) { // 認証済みの場合
+            // タスクを更新
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+        }
+            // トップページへリダイレクトさせる
         return redirect('/');
     }
 
@@ -133,11 +165,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        // idの値でタスクを検索して取得
-        $task = task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
-
+        
+        $task = Task::findOrFail($id);
+        
+        if (\Auth::id() === $task->user_id) { // 認証済みの場合
+            // タスクを削除
+            $task->delete();
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
